feat(safe-js): complete standard number typed-array kinds

+++ packages/safe-js/src/interp/uint8array.test.ts
@@ -79,7 +79,7 @@ it.each([
   await expect(run(source,{budget:new Budget(limits)})).rejects.toMatchObject({code:"budgetExceeded",budget});
 });
 
-it.each([undefined,"Float64Array","constructor","__proto__",1])("rejects unregistered snapshot type %j", arrayType => {
+it.each([undefined,"BigInt64Array","constructor","__proto__",1])("rejects unregistered snapshot type %j", arrayType => {
   expect(()=>decodeTypedArrayStorage({kind:"typedarray",arrayType,bytes:[0],byteOffset:0,length:1},()=>undefined)).toThrow(TypeError);
 });
